refactor: rename _potential to _src
