+++ src/galax/potential/_src/builtin/__init__.py
