+++ src/galax/potential/_src/builtin/bars.py
@@ -17,10 +17,10 @@ from unxt import AbstractUnitSystem, Quantity, unitsystem
 from xmmutablemap import ImmutableMap
 
 import galax.typing as gt
-from galax.potential._potential.base import default_constants
-from galax.potential._potential.core import AbstractPotential
-from galax.potential._potential.params.core import AbstractParameter
-from galax.potential._potential.params.field import ParameterField
+from galax.potential._src.base import default_constants
+from galax.potential._src.core import AbstractPotential
+from galax.potential._src.params.core import AbstractParameter
+from galax.potential._src.params.field import ParameterField
 from galax.utils._jax import vectorize_method
 
 # -------------------------------------------------------------------
