+++ src/galax/potential/_src/composite.py
