+++ src/galax/potential/_src/core.py
