+++ src/galax/potential/_src/frame.py
