+++ src/galax/potential/_src/params/attr.py
