+++ src/galax/potential/_src/params/core.py
