+++ src/galax/potential/_src/plot.py
